修复bug：代码查重查到自己

+++ judge/cpp/judge.cpp
@@ -596,7 +596,7 @@ void sim(char *ac_dir)
             continue;
         char sid[11]={0};
         strncpy(sid, dirfile->d_name, strstr(dirfile->d_name,".")-dirfile->d_name);
-        if(solution_result(sid)!=OJ_AC)  //该代码曾被重判，且没有AC，不能作为查重依据
+        if(solution.id>atoi(sid) && solution_result(sid)!=OJ_AC)  //该代码曾被重判，且没有AC，不能作为查重依据
         {
             system_cmd("rm -rf %s/%s",ac_dir,dirfile->d_name); //删除
             continue;
